SEISMIC_CODING/new/requirements.txt
numpy
pyyaml
matplotlib
torch
tqdm
